django_app/requirements.txt
Django==4.2.21
django-auth-ldap==4.6.0
python-ldap==3.4.3
gunicorn==21.2.0
python-dotenv==1.0.0 